Extract toast message args type in global ts
